Moved around files and updated CollaborativeScheduling to reflect changes in entities
