+++ src/main/java/use_case_collaborative_scheduling/CollaborativeSchedulingResponseModel.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 public class CollaborativeSchedulingResponseModel {
 
     private final ArrayList<ArrayList<LocalDateTime>> availableTimes;
+    ArrayList<String> scheduleString;
 
     public CollaborativeSchedulingResponseModel(ArrayList<ArrayList<LocalDateTime>> availableTimes) {
         this.availableTimes = availableTimes;
@@ -13,5 +14,12 @@ public class CollaborativeSchedulingResponseModel {
 
     public ArrayList<ArrayList<LocalDateTime>> getAvailableTimes() {
         return availableTimes;
+    }
+
+    public void setScheduleString(ArrayList<String> scheduleString) {
+        this.scheduleString = scheduleString;
+    }
+    public ArrayList<String> getScheduleString() {
+        return this.scheduleString;
     }
 }
